requirements.txt
boto3>=1.14.39
s3fs>=0.4.2
python-slugify>=4.0.1
click>=7.1.2
loguru>=0.5.1